Java program to check
Java Program to check if the number of digits in the square of a number, when divided in two parts, such that the second part contains equal number of digits as the number itself. When the two parts are added, a check is made if the sum is equal to the number.
Eg. 45
Square= 45*45= 2025
Part 1= 20
Part 2= 25
Sum= 45
Condition=True